+++ SleepModel/HEMNewProfileCollectionViewCell.h
@@ -1,0 +1,13 @@
+//
+//  HEMNewProfileCollectionViewCell.h
+//  Sense
+//
+//  Created by Jimmy Lu on 5/13/16.
+//  Copyright © 2016 Hello. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface HEMNewProfileCollectionViewCell : UICollectionViewCell
+
+@end