+++ SleepModel/HEMTitledTextField.h
@@ -1,0 +1,15 @@
+//
+//  HEMTitledTextField.h
+//  Sense
+//
+//  Created by Jimmy Lu on 5/12/16.
+//  Copyright © 2016 Hello. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface HEMTitledTextField : UIView
+
+@property (nonatomic, weak) IBOutlet UITextField* textField;
+
+@end